+++ src/app/app.constants.ts
@@ -1,0 +1,3 @@
+export enum AppProviderToken {
+  EXPRESS_APP_INSTANCE = 'express-app',
+}