+++ src/app/decorators/app.decorator.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata';
-import { BootstrapService } from './bootstrap.service';
-import { AppProperties } from './types/app-properties.type';
+import { BootstrapService } from '../bootstrap.service';
+import { AppProperties } from '../types/app-properties.type';
 
 export function App(properties: AppProperties): ClassDecorator {
   return function (target: any) {