+++ src/app/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './app.decorator';
+export * from './express-app-instance.decorator';