+++ src/app/index.ts
@@ -1,1 +1,2 @@
-export * from './app.decorator';
+export * from './decorators';
+export * from './lifecycle-hooks';